# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Instituto Acaz Ribeiro - Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professionals:
    get:
      operationId: listProfessionals
      tags: [professionals]
      summary: List all professionals
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of professionals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Professional"
    post:
      operationId: createProfessional
      tags: [professionals]
      summary: Create a new professional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionalBody"
      responses:
        "201":
          description: Professional created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
  /professionals/{id}:
    get:
      operationId: getProfessional
      tags: [professionals]
      summary: Get professional by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfessional
      tags: [professionals]
      summary: Update a professional
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfessionalBody"
      responses:
        "200":
          description: Updated professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
    delete:
      operationId: deleteProfessional
      tags: [professionals]
      summary: Delete a professional
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /professionals/{id}/schedule:
    get:
      operationId: getProfessionalSchedule
      tags: [professionals]
      summary: Get professional weekly schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSlot"
    put:
      operationId: updateProfessionalSchedule
      tags: [professionals]
      summary: Update professional weekly schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slots:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreateScheduleSlotBody"
              required: [slots]
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSlot"

  # Services
...
  /reports/financial:
    get:
      operationId: getFinancialReport
      tags: [reports]
      summary: Financial report by period
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: professionalId
          in: query
          schema:
            type: integer
        - name: serviceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financial report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialReport"
  /reports/appointments:
    get:
      operationId: getAppointmentReport
      tags: [reports]
      summary: Appointment report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: professionalId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointment report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentReport"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth
...
    Professional:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        cpf:
          type: ["string", "null"]
        active:
          type: boolean
        clinicSharePercent:
          type: number
          description: Percentage going to clinic (default 30)
        professionalSharePercent:
          type: number
          description: Percentage going to professional (default 70)
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - specialty
        - cpf
        - active
        - clinicSharePercent
        - professionalSharePercent
        - createdAt
    CreateProfessionalBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialty:
          type: string
        cpf:
          type: string
        clinicSharePercent:
          type: number
          default: 30
        professionalSharePercent:
          type: number
          default: 70
      required:
        - name
        - email
    UpdateProfessionalBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialty:
          type: string
        cpf:
          type: string
        active:
          type: boolean
        clinicSharePercent:
          type: number
        professionalSharePercent:
          type: number
    ScheduleSlot:
      type: object
      properties:
        id:
          type: integer
        professionalId:
          type: integer
        dayOfWeek:
          type: integer
          description: 0=Sunday, 1=Monday, ..., 6=Saturday
        startTime:
          type: string
          description: "HH:MM"
        endTime:
          type: string
          description: "HH:MM"
      required:
        - id
        - professionalId
        - dayOfWeek
        - startTime
        - endTime
    CreateScheduleSlotBody:
      type: object
      properties:
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
      required:
        - dayOfWeek
        - startTime
        - endTime

    # Services
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        professionalId:
          type: integer
        professionalName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        contractId:
          type: ["integer", "null"]
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show, blocked]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - professionalId
        - professionalName
        - serviceId
        - serviceName
        - contractId
        - date
        - startTime
        - endTime
        - status
        - notes
        - createdAt
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        professionalId:
          type: integer
        professionalName:
          type: string
        serviceId:
          type: ["integer", "null"]
        serviceName:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        packageName:
          type: ["string", "null"]
        grossAmount:
          type: number
        paymentMethod:
          type: string
          enum: [pix, credit, debit, cash]
        gateFee:
          type: number
          description: Payment gateway fee
        netAmount:
          type: number
          description: Gross minus gateway fee
        clinicAmount:
          type: number
          description: Clinic's share
        professionalAmount:
          type: number
          description: Professional's share
        appointmentId:
          type: ["integer", "null"]
        contractId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - professionalId
        - professionalName
        - serviceId
        - serviceName
        - packageId
        - packageName
        - grossAmount
        - paymentMethod
        - gateFee
        - netAmount
        - clinicAmount
        - professionalAmount
        - appointmentId
        - contractId
        - notes
        - createdAt
...
    ProfessionalStat:
      type: object
      properties:
        professionalId:
          type: integer
        professionalName:
          type: string
        appointments:
          type: integer
        revenue:
          type: number
        professionalEarnings:
          type: number
        clinicEarnings:
          type: number
      required:
        - professionalId
        - professionalName
        - appointments
        - revenue
        - professionalEarnings
        - clinicEarnings

    # Reports
    FinancialReport:
      type: object
      properties:
        totalGross:
          type: number
        totalFees:
          type: number
        totalNet:
          type: number
        totalClinic:
          type: number
        totalProfessionals:
          type: number
        byPaymentMethod:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMethodBreakdown"
        byProfessional:
          type: array
          items:
            $ref: "#/components/schemas/ProfessionalStat"
        byService:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBreakdown"
        salesData:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
      required:
        - totalGross
        - totalFees
        - totalNet
        - totalClinic
        - totalProfessionals
        - byPaymentMethod
        - byProfessional
        - byService
        - salesData
    PaymentMethodBreakdown:
      type: object
      properties:
        method:
          type: string
        count:
          type: integer
        total:
          type: number
      required:
        - method
        - count
        - total
    ServiceBreakdown:
      type: object
      properties:
        serviceId:
          type: integer
        serviceName:
          type: string
        count:
          type: integer
        total:
          type: number
      required:
        - serviceId
        - serviceName
        - count
        - total
    AppointmentReport:
      type: object
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusBreakdown"
        byProfessional:
          type: array
          items:
            $ref: "#/components/schemas/ProfessionalAppointmentStat"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required:
        - total
        - byStatus
        - byProfessional
        - appointments
    StatusBreakdown:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count
    ProfessionalAppointmentStat:
      type: object
      properties:
        professionalId:
          type: integer
        professionalName:
          type: string
        total:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
      required:
        - professionalId
        - professionalName
        - total
        - completed
        - cancelled